% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolrgrowthlearning.R, R/dolrwagehour.R
\name{dol_gol}
\alias{dol_gol}
\alias{dol_whd}
\title{dol_gol}
\usage{
dol_gol(dataset = 1, sheet = 1, key = pkg.env$curr.key)

dol_whd(dataset = 1, sheet = 1, key = pkg.env$curr.key)
}
\arguments{
\item{dataset}{There are currently 3 datasets to choose from. Specify the one you want with a number from 1-3.
\enumerate{
\item Farm Labor Contractor and Farm Labor
\item Contractor Employee
\item Wage and Hour Publication System
\item WHD Compliance
}}

\item{sheet}{For each dataset chosen, there are sub sheets within them. You must specify which sub-sheet you want.}

\item{key}{The API key for this user. You must go to the DOL and sign up for an account and request for an API key.}
}
\value{
A dataframe

A dataframe
}
\description{
This function queries the US Department of Labor Expanding Growth Opportunities and Learning Datasets. The datasets currently use V1 of DOL API.

This function queries the US Department of Labor Wage and Hour Divisions Datasets. The datasets currently use V2 of DOL API.
}
\keyword{Divisions}
\keyword{Expanding}
\keyword{Growth}
\keyword{Hour}
\keyword{Learning}
\keyword{Opportunities}
\keyword{Wage}
\keyword{and}
