% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolrhealthsafety.R
\name{dol_hsd}
\alias{dol_hsd}
\title{dol_hsd}
\usage{
dol_hsd(dataset = 1, sheet = 1, key = Sys.getenv("DOL_KEY"))
}
\arguments{
\item{dataset}{There are currently 13 datasets to choose from. Specify the one you want with a number from 1-13.
\enumerate{
\item Injuries and Illness
\item Gulf Oil Spill
\item DOL OSHA Compliance
\item MSHA Employment Production
\item Mine Violation
\item Fatal Occupational Injuries
\item DOL OSHA Enforcement V2
\item Auto Workers
\item Safety and Health
\item Full Mine Information
\item Basic Mine Information
\item Mine Accident Injuries
\item Mine Inspections
}}

\item{sheet}{For each dataset chosen, there are sub sheets within them. You must specify which sub-sheet you want. For example,
if you chose Gulf Oil Spill, there are the subsheets, 1) OSHA_Direct_Read_Sampling Table as well as 2) OSHA_NOISE_REPORT Table.}

\item{key}{The API key for this user. You must go to the DOL and sign up for an account and request for an API key.}
}
\value{
A dataframe
}
\description{
This function queries the US Department of Labor Health and Safety Datasets. The datasets currently use V1 of DOL API.
}
\keyword{Health}
\keyword{Mining}
\keyword{Safety,}
\keyword{and}
