% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolrkeys.R
\name{dolgetkey}
\alias{dolgetkey}
\title{dolgetkey}
\usage{
dolgetkey()
}
\value{
A string. NULL if empty.
}
\description{
The function will return the key set by Set.sysenv(). Returns NULL if no key was set.
}
\details{
dolgetkey
}
