% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolr.R
\name{dolsetkey}
\alias{dolsetkey}
\title{dolsetkey}
\usage{
dolsetkey(key)
}
\arguments{
\item{key}{The token given by DOL here https://devtools.dol.gov/developer}
}
\description{
The function will save the key into your library download. You will need to
update this when you call library(dolr) again. You can also set a new key by using this function with a new one.
}
\details{
dolsetkey
}
