% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolrdatasetfinder.R
\name{guess_dataset_sheet}
\alias{guess_dataset_sheet}
\title{guess_dataset_sheet}
\usage{
guess_dataset_sheet(dataset, sheet)
}
\arguments{
\item{dataset}{A list of dataset are shown at https://developer.dol.gov/dataset/}

\item{sheet}{For each dataset chosen, there are sub sheets within them.
You must specify which sub-sheet you want. Look up the corresponding sheet
at https://developer.dol.gov/dataset/.}
}
\value{
The guessed dataset and sheet.
}
\description{
This function determines what dataset/sheet the user likely selected
from a list of available datasets and sheet in this library.
}
