% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolrapione.R
\name{query_API_1}
\alias{query_API_1}
\title{query_API_1}
\usage{
query_API_1(data, sheet, key)
}
\arguments{
\item{data}{The subpath where the variable is located.}

\item{sheet}{For each dataset chosen, there are sub sheets within them.
You must specify which sub-sheet you want.
Look up the corresponding sheet at https://developer.dol.gov/dataset/.}

\item{key}{The API key for this user. You must go to the DOL and sign up
for an account and request for an API key.}
}
\value{
A dataframe
}
\description{
Helper function for query the DOL V1 API.
}
