% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dolr.R
\name{dolgetkey}
\alias{dolgetkey}
\title{dolgetkey}
\usage{
dolgetkey()
}
\value{
A string vector set by dolsetkey(). NULL if empty.
}
\description{
The function will return the key set by dolsetkey(). Returns NULL if no key was set.
}
\details{
dolgetkey
}
